pyproject.toml
[project]
name = "notebooklm-doc-sync"
version = "0.1.0"
description = "Sincronizador de documentación para NotebookLM"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "trafilatura>=1.12.0",
    "PyYAML>=6.0.1",
    "selectolax>=0.3.21",
    "click>=8.1.7",
    "pydantic>=2.0.0",
    "markdownify>=0.14.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
doc-sync = "doc_sync.cli:main"

[tool.setuptools]
packages = ["doc_sync"]
